# mobile/openapi.json
{
 "components": {
  "schemas": {
...
   "AskRequest": {
    "properties": {
     "question": {
      "maxLength": 500,
      "minLength": 1,
      "title": "Question",
      "type": "string"
     }
    },
    "required": [
     "question"
    ],
    "title": "AskRequest",
    "type": "object"
   },
   "AskResponse": {
    "properties": {
     "answer": {
      "title": "Answer",
      "type": "string"
     },
     "cached": {
      "default": false,
      "title": "Cached",
      "type": "boolean"
     },
     "planner": {
      "title": "Planner",
      "type": "string"
     },
     "sources": {
      "items": {
       "$ref": "#/components/schemas/Source"
      },
      "title": "Sources",
      "type": "array"
     },
     "steps": {
      "items": {
       "$ref": "#/components/schemas/PlanStepOut"
      },
      "title": "Steps",
      "type": "array"
     },
     "usage": {
      "additionalProperties": {
       "type": "integer"
      },
      "title": "Usage",
      "type": "object"
     },
     "views": {
      "items": {
       "discriminator": {
        "mapping": {
         "build_proposal": "#/components/schemas/BuildProposalView",
         "candidates": "#/components/schemas/CandidatesView",
         "damage": "#/components/schemas/DamageView",
         "duel": "#/components/schemas/DuelView",
         "learn_check": "#/components/schemas/LearnCheckView",
         "learners": "#/components/schemas/LearnersView",
         "learnset": "#/components/schemas/LearnsetView",
         "member_added": "#/components/schemas/MemberAddedView",
         "move_list": "#/components/schemas/MoveListView",
         "pokemon_list": "#/components/schemas/PokemonListView",
         "ranking": "#/components/schemas/RankingView",
         "set_edit": "#/components/schemas/SetEditView",
         "survive": "#/components/schemas/SurviveView",
         "type_chart": "#/components/schemas/TypeChartView"
        },
        "propertyName": "kind"
       },
       "oneOf": [
        {
         "$ref": "#/components/schemas/RankingView"
        },
        {
         "$ref": "#/components/schemas/PokemonListView"
        },
        {
         "$ref": "#/components/schemas/TypeChartView"
        },
        {
         "$ref": "#/components/schemas/MoveListView"
        },
        {
         "$ref": "#/components/schemas/LearnsetView"
        },
        {
         "$ref": "#/components/schemas/LearnersView"
        },
        {
         "$ref": "#/components/schemas/LearnCheckView"
        },
        {
         "$ref": "#/components/schemas/CandidatesView"
        },
        {
         "$ref": "#/components/schemas/SetEditView"
        },
        {
         "$ref": "#/components/schemas/MemberAddedView"
        },
        {
         "$ref": "#/components/schemas/DuelView"
        },
        {
         "$ref": "#/components/schemas/DamageView"
        },
        {
         "$ref": "#/components/schemas/SurviveView"
        },
        {
         "$ref": "#/components/schemas/BuildProposalView"
        }
       ]
      },
      "title": "Views",
      "type": "array"
     }
    },
    "required": [
     "answer",
     "planner",
     "sources"
    ],
    "title": "AskResponse",
    "type": "object"
   },
...
   "BuildProposalView": {
    "properties": {
     "build": {
      "$ref": "#/components/schemas/BuildSuggestion"
     },
     "chunk_refs": {
      "items": {
       "type": "integer"
      },
      "title": "Chunk Refs",
      "type": "array"
     },
     "kind": {
      "const": "build_proposal",
      "default": "build_proposal",
      "title": "Kind",
      "type": "string"
     },
     "pokemon": {
      "title": "Pokemon",
      "type": "string"
     },
     "slot": {
      "title": "Slot",
      "type": "integer"
     },
     "step": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Step"
     }
    },
    "required": [
     "slot",
     "pokemon",
     "build"
    ],
    "title": "BuildProposalView",
    "type": "object"
   },
...
   "CalcField": {
    "properties": {
     "burn": {
      "default": false,
      "title": "Burn",
      "type": "boolean"
     },
     "crit": {
      "default": false,
      "title": "Crit",
      "type": "boolean"
     },
     "friend_guard": {
      "default": false,
      "title": "Friend Guard",
      "type": "boolean"
     },
     "lightscreen": {
      "default": false,
      "title": "Lightscreen",
      "type": "boolean"
     },
     "reflect": {
      "default": false,
      "title": "Reflect",
      "type": "boolean"
     },
     "terrain": {
      "default": "None",
      "title": "Terrain",
      "type": "string"
     },
     "weather": {
      "default": "None",
      "title": "Weather",
      "type": "string"
     }
    },
    "title": "CalcField",
    "type": "object"
   },
...
   "CalcSlot": {
    "description": "One filled calculator slot: 0–1 are the user's side, 2–3 the opponent's.",
    "properties": {
     "ability": {
      "default": "None",
      "title": "Ability",
      "type": "string"
     },
     "aim": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Aim"
     },
     "evs": {
      "additionalProperties": {
       "type": "integer"
      },
      "title": "Evs",
      "type": "object"
     },
     "form_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Form Id"
     },
     "hp": {
      "default": 100,
      "maximum": 100.0,
      "minimum": 0.0,
      "title": "Hp",
      "type": "number"
     },
     "item": {
      "default": "None",
      "title": "Item",
      "type": "string"
     },
     "ivs": {
      "additionalProperties": {
       "type": "integer"
      },
      "title": "Ivs",
      "type": "object"
     },
     "move": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Move"
     },
     "name": {
      "default": "",
      "title": "Name",
      "type": "string"
     },
     "nature": {
      "default": "Hardy",
      "title": "Nature",
      "type": "string"
     },
     "pokemon_id": {
      "title": "Pokemon Id",
      "type": "integer"
     },
     "slot": {
      "maximum": 3.0,
      "minimum": 0.0,
      "title": "Slot",
      "type": "integer"
     }
    },
    "required": [
     "slot",
     "pokemon_id"
    ],
    "title": "CalcSlot",
    "type": "object"
   },
...
   "CalcTurnOut": {
    "properties": {
     "aims": {
      "additionalProperties": {
       "anyOf": [
        {
         "type": "integer"
        },
        {
         "type": "null"
        }
       ]
      },
      "title": "Aims",
      "type": "object"
     },
     "hp": {
      "items": {
       "$ref": "#/components/schemas/CalcHpOut"
      },
      "title": "Hp",
      "type": "array"
     },
     "moves": {
      "items": {
       "$ref": "#/components/schemas/CalcMoveOut"
      },
      "title": "Moves",
      "type": "array"
     },
     "order": {
      "items": {
       "$ref": "#/components/schemas/CalcOrderOut"
      },
      "title": "Order",
      "type": "array"
     },
     "steps": {
      "items": {
       "$ref": "#/components/schemas/CalcStepOut"
      },
      "title": "Steps",
      "type": "array"
     }
    },
    "required": [
     "order",
     "steps",
     "hp",
     "aims",
     "moves"
    ],
    "title": "CalcTurnOut",
    "type": "object"
   },
   "CalcTurnRequest": {
    "description": "The calculator's state, to play one turn (``POST /api/calc/turn``).",
    "properties": {
     "doubles": {
      "default": false,
      "title": "Doubles",
      "type": "boolean"
     },
     "field": {
      "$ref": "#/components/schemas/CalcField"
     },
     "level": {
      "default": 100,
      "maximum": 100.0,
      "minimum": 1.0,
      "title": "Level",
      "type": "integer"
     },
     "slots": {
      "items": {
       "$ref": "#/components/schemas/CalcSlot"
      },
      "maxItems": 4,
      "title": "Slots",
      "type": "array"
     }
    },
    "title": "CalcTurnRequest",
    "type": "object"
   },
...
   "CandidatesView": {
    "description": "Pokémon the coach suggests adding; each card adds only when clicked.",
    "properties": {
     "candidates": {
      "items": {
       "$ref": "#/components/schemas/Candidate"
      },
      "title": "Candidates",
      "type": "array"
     },
     "chunk_refs": {
      "items": {
       "type": "integer"
      },
      "title": "Chunk Refs",
      "type": "array"
     },
     "kind": {
      "const": "candidates",
      "default": "candidates",
      "title": "Kind",
      "type": "string"
     },
     "members": {
      "items": {
       "$ref": "#/components/schemas/SlotRef"
      },
      "title": "Members",
      "type": "array"
     },
     "step": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Step"
     },
     "team_full": {
      "default": false,
      "title": "Team Full",
      "type": "boolean"
     },
     "team_id": {
      "title": "Team Id",
      "type": "integer"
     }
    },
    "required": [
     "team_id",
     "candidates"
    ],
    "title": "CandidatesView",
    "type": "object"
   },
...
   "DamageView": {
    "properties": {
     "apply": {
      "items": {
       "$ref": "#/components/schemas/CalcApply"
      },
      "title": "Apply",
      "type": "array"
     },
     "attacker": {
      "$ref": "#/components/schemas/CalcRef"
     },
     "changes": {
      "items": {
       "additionalProperties": true,
       "type": "object"
      },
      "title": "Changes",
      "type": "array"
     },
     "chunk_refs": {
      "items": {
       "type": "integer"
      },
      "title": "Chunk Refs",
      "type": "array"
     },
     "current": {
      "$ref": "#/components/schemas/HitRange"
     },
     "defender": {
      "$ref": "#/components/schemas/CalcRef"
     },
     "kind": {
      "const": "damage",
      "default": "damage",
      "title": "Kind",
      "type": "string"
     },
     "move": {
      "$ref": "#/components/schemas/MoveRow"
     },
     "step": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Step"
     },
     "whatif": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/HitRange"
       },
       {
        "type": "null"
       }
      ]
     }
    },
    "required": [
     "attacker",
     "defender",
     "move",
     "current"
    ],
    "title": "DamageView",
    "type": "object"
   },
...
   "DuelView": {
    "description": "One pairing played out turn by turn by the deterministic duel engine.",
    "properties": {
     "chunk_refs": {
      "items": {
       "type": "integer"
      },
      "title": "Chunk Refs",
      "type": "array"
     },
     "duel": {
      "$ref": "#/components/schemas/DuelOut"
     },
     "kind": {
      "const": "duel",
      "default": "duel",
      "title": "Kind",
      "type": "string"
     },
     "opponent_id": {
      "title": "Opponent Id",
      "type": "integer"
     },
     "step": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Step"
     },
     "team_id": {
      "title": "Team Id",
      "type": "integer"
     }
    },
    "required": [
     "team_id",
     "opponent_id",
     "duel"
    ],
    "title": "DuelView",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "LearnCheckView": {
    "description": "Can this Pokémon learn this move? Yes (with how) or no.",
    "properties": {
     "absent": {
      "default": false,
      "title": "Absent",
      "type": "boolean"
     },
     "chunk_refs": {
      "items": {
       "type": "integer"
      },
      "title": "Chunk Refs",
      "type": "array"
     },
     "game": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Game"
     },
     "how": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "How"
     },
     "kind": {
      "const": "learn_check",
      "default": "learn_check",
      "title": "Kind",
      "type": "string"
     },
     "max_level": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Level"
     },
     "method": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Method"
     },
     "move": {
      "$ref": "#/components/schemas/MoveRow"
     },
     "ok": {
      "title": "Ok",
      "type": "boolean"
     },
     "pokemon": {
      "$ref": "#/components/schemas/PokemonCard"
     },
     "step": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Step"
     }
    },
    "required": [
     "pokemon",
     "move",
     "ok"
    ],
    "title": "LearnCheckView",
    "type": "object"
   },
...
   "LearnersView": {
    "description": "Who learns a move: how many, by which method, and the standout users.",
    "properties": {
     "by_method": {
      "additionalProperties": {
       "type": "integer"
      },
      "title": "By Method",
      "type": "object"
     },
     "chunk_refs": {
      "items": {
       "type": "integer"
      },
      "title": "Chunk Refs",
      "type": "array"
     },
     "game": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Game"
     },
     "kind": {
      "const": "learners",
      "default": "learners",
      "title": "Kind",
      "type": "string"
     },
     "max_level": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Max Level"
     },
     "method": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Method"
     },
     "move": {
      "$ref": "#/components/schemas/MoveRow"
     },
     "rows": {
      "items": {
       "$ref": "#/components/schemas/PokemonCard"
      },
      "title": "Rows",
      "type": "array"
     },
     "scope": {
      "items": {
       "type": "string"
      },
      "title": "Scope",
      "type": "array"
     },
     "step": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Step"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "move",
     "total",
     "rows"
    ],
    "title": "LearnersView",
    "type": "object"
   },
...
   "LearnsetView": {
    "description": "A Pokémon's moves, grouped by how it learns them.",
    "properties": {
     "chunk_refs": {
      "items": {
       "type": "integer"
      },
      "title": "Chunk Refs",
      "type": "array"
     },
     "game": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Game"
     },
     "groups": {
      "items": {
       "$ref": "#/components/schemas/LearnGroup"
      },
      "title": "Groups",
      "type": "array"
     },
     "kind": {
      "const": "learnset",
      "default": "learnset",
      "title": "Kind",
      "type": "string"
     },
     "pokemon": {
      "$ref": "#/components/schemas/PokemonCard"
     },
     "step": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Step"
     }
    },
    "required": [
     "pokemon",
     "groups"
    ],
    "title": "LearnsetView",
    "type": "object"
   },
...
   "MemberAddedView": {
    "description": "An explicit \"add X\": what was added where (Undo clears the slot), or why not.",
    "properties": {
     "added": {
      "title": "Added",
      "type": "boolean"
     },
     "card": {
      "$ref": "#/components/schemas/PokemonCard"
     },
     "chunk_refs": {
      "items": {
       "type": "integer"
      },
      "title": "Chunk Refs",
      "type": "array"
     },
     "kind": {
      "const": "member_added",
      "default": "member_added",
      "title": "Kind",
      "type": "string"
     },
     "message": {
      "title": "Message",
      "type": "string"
     },
     "slot": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Slot"
     },
     "step": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Step"
     },
     "team_id": {
      "title": "Team Id",
      "type": "integer"
     }
    },
    "required": [
     "team_id",
     "added",
     "card",
     "message"
    ],
    "title": "MemberAddedView",
    "type": "object"
   },
...
   "MoveListView": {
    "properties": {
     "chunk_refs": {
      "items": {
       "type": "integer"
      },
      "title": "Chunk Refs",
      "type": "array"
     },
     "kind": {
      "const": "move_list",
      "default": "move_list",
      "title": "Kind",
      "type": "string"
     },
     "moves": {
      "items": {
       "$ref": "#/components/schemas/MoveRow"
      },
      "title": "Moves",
      "type": "array"
     },
     "step": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Step"
     }
    },
    "required": [
     "moves"
    ],
    "title": "MoveListView",
    "type": "object"
   },
...
   "PlanStepOut": {
    "description": "One executed plan step with its final state.",
    "properties": {
     "args": {
      "additionalProperties": true,
      "title": "Args",
      "type": "object"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "replan": {
      "default": false,
      "title": "Replan",
      "type": "boolean"
     },
     "state": {
      "default": "pending",
      "title": "State",
      "type": "string"
     },
     "summary": {
      "default": "",
      "title": "Summary",
      "type": "string"
     },
     "tool": {
      "title": "Tool",
      "type": "string"
     },
     "why": {
      "default": "",
      "title": "Why",
      "type": "string"
     }
    },
    "required": [
     "id",
     "tool"
    ],
    "title": "PlanStepOut",
    "type": "object"
   },
...
   "PokemonListView": {
    "properties": {
     "cards": {
      "items": {
       "$ref": "#/components/schemas/PokemonCard"
      },
      "title": "Cards",
      "type": "array"
     },
     "chunk_refs": {
      "items": {
       "type": "integer"
      },
      "title": "Chunk Refs",
      "type": "array"
     },
     "kind": {
      "const": "pokemon_list",
      "default": "pokemon_list",
      "title": "Kind",
      "type": "string"
     },
     "step": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Step"
     },
     "title": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Title"
     }
    },
    "required": [
     "cards"
    ],
    "title": "PokemonListView",
    "type": "object"
   },
...
   "RankingView": {
    "properties": {
     "chunk_refs": {
      "items": {
       "type": "integer"
      },
      "title": "Chunk Refs",
      "type": "array"
     },
     "kind": {
      "const": "ranking",
      "default": "ranking",
      "title": "Kind",
      "type": "string"
     },
     "order": {
      "default": "desc",
      "enum": [
       "asc",
       "desc"
      ],
      "title": "Order",
      "type": "string"
     },
     "rows": {
      "items": {
       "$ref": "#/components/schemas/RankingRow"
      },
      "title": "Rows",
      "type": "array"
     },
     "stat": {
      "title": "Stat",
      "type": "string"
     },
     "step": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Step"
     },
     "total": {
      "title": "Total",
      "type": "integer"
     }
    },
    "required": [
     "stat",
     "total",
     "rows"
    ],
    "title": "RankingView",
    "type": "object"
   },
...
   "SetEditView": {
    "description": "A proposed set for one member, shown was → now; saved only on Apply.",
    "properties": {
     "after": {
      "$ref": "#/components/schemas/BuildSet"
     },
     "before": {
      "$ref": "#/components/schemas/BuildSet"
     },
     "chunk_refs": {
      "items": {
       "type": "integer"
      },
      "title": "Chunk Refs",
      "type": "array"
     },
     "fields": {
      "additionalProperties": true,
      "title": "Fields",
      "type": "object"
     },
     "kind": {
      "const": "set_edit",
      "default": "set_edit",
      "title": "Kind",
      "type": "string"
     },
     "member": {
      "additionalProperties": true,
      "title": "Member",
      "type": "object"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "side": {
      "enum": [
       "ours",
       "theirs"
      ],
      "title": "Side",
      "type": "string"
     },
     "slot": {
      "title": "Slot",
      "type": "integer"
     },
     "sprite_url": {
      "default": "",
      "title": "Sprite Url",
      "type": "string"
     },
     "step": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Step"
     },
     "team_id": {
      "title": "Team Id",
      "type": "integer"
     },
     "why": {
      "default": "",
      "title": "Why",
      "type": "string"
     }
    },
    "required": [
     "side",
     "team_id",
     "slot",
     "name",
     "before",
     "after"
    ],
    "title": "SetEditView",
    "type": "object"
   },
...
   "Source": {
    "properties": {
     "chunk_type": {
      "title": "Chunk Type",
      "type": "string"
     },
     "dex_number": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Dex Number"
     },
     "n": {
      "title": "N",
      "type": "integer"
     },
     "pokemon_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pokemon Id"
     },
     "pokemon_name": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pokemon Name"
     },
     "score": {
      "title": "Score",
      "type": "number"
     },
     "snippet": {
      "title": "Snippet",
      "type": "string"
     },
     "source_ref": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Source Ref"
     },
     "step": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Step"
     },
     "step_index": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Step Index"
     }
    },
    "required": [
     "n",
     "chunk_type",
     "snippet",
     "score"
    ],
    "title": "Source",
    "type": "object"
   },
...
   "SurviveView": {
    "properties": {
     "already": {
      "default": false,
      "title": "Already",
      "type": "boolean"
     },
     "apply": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/CalcApply"
       },
       {
        "type": "null"
       }
      ]
     },
     "attacker": {
      "$ref": "#/components/schemas/CalcRef"
     },
     "chunk_refs": {
      "items": {
       "type": "integer"
      },
      "title": "Chunk Refs",
      "type": "array"
     },
     "current": {
      "$ref": "#/components/schemas/HitRange"
     },
     "defender": {
      "$ref": "#/components/schemas/CalcRef"
     },
     "hp_ev": {
      "title": "Hp Ev",
      "type": "integer"
     },
     "kind": {
      "const": "survive",
      "default": "survive",
      "title": "Kind",
      "type": "string"
     },
     "move": {
      "$ref": "#/components/schemas/MoveRow"
     },
     "nature": {
      "title": "Nature",
      "type": "string"
     },
     "nature_changed": {
      "title": "Nature Changed",
      "type": "boolean"
     },
     "range": {
      "$ref": "#/components/schemas/HitRange"
     },
     "stat": {
      "enum": [
       "def",
       "spd"
      ],
      "title": "Stat",
      "type": "string"
     },
     "stat_ev": {
      "title": "Stat Ev",
      "type": "integer"
     },
     "step": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Step"
     },
     "survives": {
      "title": "Survives",
      "type": "boolean"
     }
    },
    "required": [
     "defender",
     "attacker",
     "move",
     "survives",
     "stat",
     "hp_ev",
     "stat_ev",
     "nature",
     "nature_changed",
     "range",
     "current"
    ],
    "title": "SurviveView",
    "type": "object"
   },
...
   "TypeChartView": {
    "description": "How one type (or a dual typing) fares defensively.",
    "properties": {
     "chunk_refs": {
      "items": {
       "type": "integer"
      },
      "title": "Chunk Refs",
      "type": "array"
     },
     "immune": {
      "items": {
       "type": "string"
      },
      "title": "Immune",
      "type": "array"
     },
     "kind": {
      "const": "type_chart",
      "default": "type_chart",
      "title": "Kind",
      "type": "string"
     },
     "resist_half": {
      "items": {
       "type": "string"
      },
      "title": "Resist Half",
      "type": "array"
     },
     "resist_quarter": {
      "items": {
       "type": "string"
      },
      "title": "Resist Quarter",
      "type": "array"
     },
     "step": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Step"
     },
     "strong_against": {
      "items": {
       "type": "string"
      },
      "title": "Strong Against",
      "type": "array"
     },
     "types": {
      "items": {
       "type": "string"
      },
      "title": "Types",
      "type": "array"
     },
     "weak_2x": {
      "items": {
       "type": "string"
      },
      "title": "Weak 2X",
      "type": "array"
     },
     "weak_4x": {
      "items": {
       "type": "string"
      },
      "title": "Weak 4X",
      "type": "array"
     }
    },
    "required": [
     "types"
    ],
    "title": "TypeChartView",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  }
 },
 "info": {
  "description": "Backend for the premium Pokédex + personalized RAG project.",
  "title": "pokérag API",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/ask": {
   "post": {
    "operationId": "ask_api_ask_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AskRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/AskResponse"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Ask",
    "tags": [
     "ask"
    ]
   }
  },
...
  "/api/calc/turn": {
   "post": {
    "description": "Play one calculator turn: order, targeting, HP carried over, Focus Sash, KO calls.\n\nThe same turn the website shows (``lib/calcTurn``), with types, stats and moves re-read\nfrom the DB. Pure code, no LLM.",
    "operationId": "calc_turn_route_api_calc_turn_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CalcTurnRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CalcTurnOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Calc Turn Route",
    "tags": [
     "calc"
    ]
   }
  },
...
 }
}